requirements.txt
aiohttp==3.9.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
attrs==23.2.0
blinker==1.7.0
Bootstrap-Flask==2.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.0
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
soupsieve==2.5
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
WTForms==3.1.2
yarl==1.9.4
